Remove Icon component
