--- packages/harmony/src/components/icon/index.ts
@@ -1,1 +1,0 @@
-export * from './Icon'
