+++ packages/harmony/src/foundations/theme/theme.ts
@@ -1,4 +1,3 @@
-import { iconSizes } from '../../components/icon'
 import type { ColorTheme } from '../color'
 import { colorTheme } from '../color'
 import { CornerRadius, cornerRadius } from '../corner-radius'
@@ -9,6 +8,15 @@ import { typography } from '../typography'
 import type { Typography } from '../typography'
 
 import type { Theme } from './types'
+
+const iconSizes = {
+  xs: 14,
+  s: 16,
+  m: 20,
+  l: 24,
+  xl: 30,
+  '2xl': 32
+}
 
 const commonFoundations = {
   shadows,
