+++ packages/mobile/src/app/index.ts
@@ -1,3 +1,4 @@
-export * from './App'
+// export * from './App'
+export { App } from './Tmp'
 // Uncomment to run app in storybook mode
 // export { default as App } from '../../.storybook'
